Update READMe.md
For the sake of the missing column you are working on mentioned in the **NOTES** section, I suggest adding a distinction or column to check-off all the "non-custodial" exchanges, where they never take custody, yet still are centralized in the sense of some admins/persons having the ability to shut the site down (i.e. the site or software isn't fully decentralized yet) and is still controlled by certain users. I added some comments into the notes section for review, hope that helps, and great work on compiling this list!